pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "finops-sim"
version = "0.2.0"
description = "FinOps simulation problem generator for training/evaluating FinOps AI"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "eval_type_backport>=0.1.0; python_version < '3.10'",
]

[project.optional-dependencies]
llm = ["anthropic>=0.30"]
scoring = ["jsonschema>=4.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "faker>=25.0",
    "jsonschema>=4.0",
]

[project.scripts]
finops-sim = "finops_sim.cli:main"
finops-study = "finops_study.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
